requirements.txt
customtkinter
Pillow
sqlite3
pygame
numpy